Seperate Zendesk normal service + remove redundancy code
